requirements.txt
Django>=3.0,<4.0
djangorestframework==3.11.1
django-rq==2.3.2
psycopg2-binary>=2.8
pytest==6.0.1
pytest-django==3.9.0
pytest-mock==3.3.1
